+++ launch/zed_camera.launch
@@ -1,7 +1,10 @@
 <launch>
   <arg name="svo_file"/>
   <include file="$(find zed_wrapper)/launch/zed_tf.launch" />
-  <node name="zed_wrapper_node" pkg="zed_wrapper" type="zed_wrapper_node" args="$(arg svo_file)" output="screen">
+  <node name="zed_wrapper_node" pkg="zed_wrapper" type="zed_wrapper_node" output="screen">
+
+    <!-- SVO file path -->
+    <param name="svo_filepath" value="$(arg svo_file)" />
 
     <!-- ZED parameters -->
     <param name="resolution"            value="2" />
