+++ package.xml
@@ -20,6 +20,7 @@
   <build_depend>dynamic_reconfigure</build_depend>
   <build_depend>pcl_conversions</build_depend>
   <build_depend>urdf</build_depend>
+  <build_depend>nodelet</build_depend>
 
   <run_depend>tf2_ros</run_depend>
   <run_depend>nav_msgs</run_depend>
@@ -30,6 +31,9 @@
   <run_depend>image_transport</run_depend>
   <run_depend>dynamic_reconfigure</run_depend>
   <run_depend>pcl_conversions</run_depend>
+  <run_depend>nodelet</run_depend>
+  <run_depend>robot_state_publisher</run_depend>
   <export>
+    <nodelet plugin="${prefix}/nodelet_plugins.xml" />
   </export>
 </package>
